requirements.txt
Django >= 1.5
south >= 0.7.6
PIL >= 1.1.7
django-annoying